#15 fix: current HP bar disappeared #20 overlay layout tweaks

+++ Assets/Scripts/UI/Data/Hero+Fields.cs
@@ -84,7 +84,7 @@ namespace Assets.Scripts.UI.Data
         }
         public List<BarInfo> BarsInfoShort => BarsInfo.GetRange(0, 2);
         public List<BarInfo> BarsInfoBattle => new() {
-            BarInfo.EmptyBarInfo(0, $"HP: {HealthCurrent}", Color.red, HealthCurrent / Health),
+            BarInfo.EmptyBarInfo(0, $"HP: {HealthCurrent}", Color.red, (float)HealthCurrent / Health),
             BarInfo.EmptyBarInfo(1, $"Speed: {Speed}", null, Speed / 10f),
         };
